refactor: moves fly to method to more appropriate module
